Add Initial Migrations

+++ AlibabaClone.Domain/Aggregates/AccountAggregates/Gender.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace AlibabaClone.Domain.Aggregates.AccountAggregates
 {
-    public class Gender : Entity<byte>
+    public class Gender : Entity<short>
     {
         public required string Title { get; set; }
     }
